Small bugfix to RPC cache

+++ src/cache/rpc_cache.go
@@ -123,7 +123,7 @@ func (cache *rpcCache) Retrieve(target *core.BuildTarget, key []byte) bool {
 	if len(req.Artifacts) == 0 {
 		return false
 	}
-	return cache.retrieveArtifacts(target, &req)
+	return cache.retrieveArtifacts(target, &req, true)
 }
 
 func (cache *rpcCache) RetrieveExtra(target *core.BuildTarget, key []byte, file string) bool {
@@ -133,10 +133,10 @@ func (cache *rpcCache) RetrieveExtra(target *core.BuildTarget, key []byte, file 
 	artifact := pb.Artifact{Package: target.Label.PackageName, Target: target.Label.Name, File: file}
 	artifacts := []*pb.Artifact{&artifact}
 	req := pb.RetrieveRequest{Hash: key, Os: runtime.GOOS, Arch: runtime.GOARCH, Artifacts: artifacts}
-	return cache.retrieveArtifacts(target, &req)
+	return cache.retrieveArtifacts(target, &req, false)
 }
 
-func (cache *rpcCache) retrieveArtifacts(target *core.BuildTarget, req *pb.RetrieveRequest) bool {
+func (cache *rpcCache) retrieveArtifacts(target *core.BuildTarget, req *pb.RetrieveRequest, remove bool) bool {
 	ctx, cancel := context.WithTimeout(context.Background(), cache.timeout)
 	defer cancel()
 	response, err := cache.client.Retrieve(ctx, req)
@@ -152,11 +152,13 @@ func (cache *rpcCache) retrieveArtifacts(target *core.BuildTarget, req *pb.Retri
 	// Remove any existing outputs first; this is important for cases where the output is a
 	// directory, because we get back individual artifacts, and we need to make sure that
 	// only the retrieved artifacts are present in the output.
-	for _, out := range target.Outputs() {
-		out := path.Join(target.OutDir(), out)
-		if err := os.RemoveAll(out); err != nil {
-			log.Error("Failed to remove artifact %s: %s", out, err)
-			return false
+	if remove {
+		for _, out := range target.Outputs() {
+			out := path.Join(target.OutDir(), out)
+			if err := os.RemoveAll(out); err != nil {
+				log.Error("Failed to remove artifact %s: %s", out, err)
+				return false
+			}
 		}
 	}
 	for _, artifact := range response.Artifacts {
